// src/types/InteractToEarn.d.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */

import BN from "bn.js";
import { ContractOptions } from "web3-eth-contract";
import { EventLog } from "web3-core";
import { EventEmitter } from "events";
import {
  Callback,
  PayableTransactionObject,
  NonPayableTransactionObject,
  BlockType,
  ContractEventLog,
  BaseContract,
} from "./types";

interface EventOptions {
  filter?: object;
  fromBlock?: BlockType;
  topics?: string[];
}

export interface InteractToEarn extends BaseContract {
  constructor(
    jsonInterface: any[],
    address?: string,
    options?: ContractOptions
  ): InteractToEarn;
  clone(): InteractToEarn;
  methods: {
    balanceReceived(): NonPayableTransactionObject<string>;

    Deposit(depositValue: number | string | BN): PayableTransactionObject<void>;

    farmBalance(): PayableTransactionObject<void>;

    withdrawBalance(
      withdrawValue: number | string | BN
    ): PayableTransactionObject<void>;

    getDepositBlock(): NonPayableTransactionObject<string>;

    currentBalance(): NonPayableTransactionObject<string>;
  };
  events: {
    allEvents(options?: EventOptions, cb?: Callback<EventLog>): EventEmitter;
  };
}